Update README.md
PLOT
Chihiro Ogino, a disaffected child, is annoyed about having to move to a new town. She is traveling with her parents in their 1996 Audi A4 Quattro to their new home. While driving to their new house, Chihiro's father attempts to follow a shortcut; they subsequently lose their way and come across a mysterious red gate and a tunnel which exits to a clock tower and leads to what appears to be an abandoned theme park, lined with seemingly empty restaurants. Finding a restaurant fully-stocked with unattended food, both parents eat the food they find there and, as a result, transform into pigs.